better way to do biguint inverse mod
