+++ primitives/src/ecc/mod.rs
@@ -1,8 +1,4 @@
-use num_bigint_dig::BigUint;
-
-use crate::bigint::{
-    check_carry_mod_to_zero::CheckCarryModToZeroSubAir, CanonicalUint, LimbConfig,
-};
+use crate::bigint::{CanonicalUint, LimbConfig};
 
 pub mod air;
 pub mod columns;
@@ -14,51 +10,4 @@ mod tests;
 pub struct EcPoint<T, C: LimbConfig> {
     pub x: CanonicalUint<T, C>,
     pub y: CanonicalUint<T, C>,
-}
-
-pub struct EccAir {
-    // e.g. secp256k1 is 2^256 - 2^32 - 977.
-    pub prime: BigUint,
-
-    // y^2 = x^3 + b. b=7 for secp256k1.
-    pub b: BigUint,
-
-    // The limb config for the EcPoint coordinates.
-    pub limb_bits: usize,
-    // Number of limbs of the prime and the coordinates.
-    pub num_limbs: usize,
-
-    // The subair to constrain big integer operations.
-    pub check_carry: CheckCarryModToZeroSubAir,
-    // Range checker decomp bits.
-    pub decomp: usize,
-}
-
-impl EccAir {
-    pub fn new(
-        prime: BigUint,
-        b: BigUint,
-        range_checker_bus: usize,
-        decomp: usize,
-        limb_bits: usize,
-        field_element_bits: usize,
-    ) -> Self {
-        let num_limbs = (prime.bits() + limb_bits - 1) / limb_bits;
-        let check_carry = CheckCarryModToZeroSubAir::new(
-            prime.clone(),
-            limb_bits,
-            range_checker_bus,
-            decomp,
-            field_element_bits,
-        );
-
-        EccAir {
-            prime,
-            b,
-            limb_bits,
-            num_limbs,
-            check_carry,
-            decomp,
-        }
-    }
 }
